first reorganisation of files into different folders
