+++ gusto/__init__.py
@@ -1,29 +1,15 @@
 
 # Start logging first, incase anything goes wrong
-from gusto.logging import *                          # noqa
+from gusto.core.logging import *                     # noqa
 set_log_handler()
 
-from gusto.active_tracers import *                   # noqa
-from gusto.common_forms import *                     # noqa
-from gusto.configuration import *                    # noqa
-from gusto.coord_transforms import *                 # noqa
-from gusto.domain import *                           # noqa
+from gusto.core import *                             # noqa
 from gusto.diagnostics import *                      # noqa
-from gusto.diffusion_methods import *                # noqa
 from gusto.equations import *                        # noqa
-from gusto.forcing import *                          # noqa
-from gusto.initialisation_tools import *             # noqa
-from gusto.io import *                               # noqa
-from gusto.labels import *                           # noqa
-from gusto.limiters import *                         # noqa
-from gusto.linear_solvers import *                   # noqa
-from gusto.meshes import *                           # noqa
-from gusto.numerical_integrator import *             # noqa
+from gusto.initialisation import *                   # noqa
 from gusto.physics import *                          # noqa
-from gusto.preconditioners import *                  # noqa
 from gusto.recovery import *                         # noqa
+from gusto.solvers import *                          # noqa
 from gusto.spatial_methods import *                  # noqa
 from gusto.time_discretisation import *              # noqa
-from gusto.timeloop import *                         # noqa
-from gusto.transport_methods import *                # noqa
-from gusto.wrappers import *                         # noqa
+from gusto.timestepping import *                     # noqa
