+++ gusto/core/coordinates.py
@@ -3,8 +3,8 @@ This file provides a coordinate object, dependent on the mesh.
 Coordinate fields are stored in specified VectorFunctionSpaces.
 """
 
-from gusto.coord_transforms import lonlatr_from_xyz, rotated_lonlatr_coords
-from gusto.logging import logger
+from gusto.core.coord_transforms import lonlatr_from_xyz, rotated_lonlatr_coords
+from gusto.core.logging import logger
 from firedrake import SpatialCoordinate, Function
 import numpy as np
 import pandas as pd
