+++ unit-tests/test_spherical_vector_transforms.py
@@ -6,7 +6,7 @@ Firedrake vector function spaces.
 import importlib
 import numpy as np
 import firedrake as fd
-from gusto.coord_transforms import *
+from gusto.core.coord_transforms import *
 import pytest
 
 tol = 1e-12
